Don't gclient config/sync an extra time for Android.
This didn't work, because it was downloading to the wrong directory;
i.e., the current directory, intead of third_party/webrtc. After running
build.sh, you would have a second checkout of webrtc in ./src, as well
as a .gclient file. The addition of "target_os = [ 'android' ]" to
third_party/webrtc/.gclient was working properly, but it would take
effect until the next run and the next "gclient sync".