Set GN arg symbol_level=0.
This is to omit more debugging information and make the output library
smaller on Android. It turns out that is_debug doesn't directly control
the inclusion of debugging symbols, it just affects the logic for
computing symbol_level.
https://chromium.googlesource.com/chromium/src/build/+/99653454ee423cc5528426f9827dc24627f95c6a/config/compiler/compiler.gni#175
On Android, the default (with is_debug=false) is 1, while on linux and
mac the default is 0. Here is where it is applied to actually set
compiler flags:
https://chromium.googlesource.com/chromium/src/build/+/99653454ee423cc5528426f9827dc24627f95c6a/config/compiler/BUILD.gn#2009

This change reduces the size of libwebrtc-android-arm-magic.a from 97M
to 45M, and that of third_party/webrtc/src/out/Release/obj/libwebrtc.a
from 77M to 38M. There is no change in libwebrtc-linux-amd64-magic.a nor
libwebrtc-linux-arm-magic.a.

Description from `gn --list=is_debug args out/Release`:
symbol_level
    How many symbols to include in the build. This affects the performance of
    the build since the symbols are large and dealing with them is slow.
      2 means regular build with symbols.
      1 means minimal symbols, usually enough for backtraces only. Symbols with
    internal linkage (static functions or those in anonymous namespaces) may not
    appear when using this level.
      0 means no symbols.
      -1 means auto-set according to debug/release and platform.